pk_project/pk_app/serializers.py
from rest_framework import serializers
from pk_app.models import Currency

class CurrencySerializer(serializers.ModelSerializer):
    class Meta:
        model = Currency
        fields = ('id', 'name', 'rate')