requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-json-rpc==0.7.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.6
jsonrpcserver==5.0.9
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.1.0
OSlash==0.6.3
psycopg2==2.9.9
pytz==2024.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1